convert journal entry
