+++ src/lib/journalEntry.ts
@@ -1,15 +1,9 @@
 import { Prisma, type JournalEntry, type PrismaClient } from '@prisma/client';
-import { z } from 'zod';
+import type { z } from 'zod';
 import { APIError } from './errors';
 import type { Repo } from './repo';
 import { matchMetricsInString, parseMetricStrings, toNum } from './utils';
-
-export const journalEntrySchema = z.object({
-  date: z.date().default(new Date()),
-  content: z.string().min(1, { message: 'Journal entry content is required' }),
-  type: z.string().default('climbing'),
-});
-export type JournalEntrySchema = typeof journalEntrySchema;
+import type { JournalEntrySchema } from './zodSchemas';
 
 export class JournalEntryRepo implements Repo<JournalEntry, Prisma.JournalEntrySelect> {
   constructor(private readonly prisma: PrismaClient) {}
